Move builder_attr types into the related *_info.rs
When there was just BuilderAttr, it being separate made some sense. Now
there is TypeBuilderAttr to join FieldBuilderAttr, keeping them together
makes little sense, so it’s a question of splitting them into their own
files, or keeping them together with their Info struct. I think keeping
them together makes sense.
