+++ src/field_info.rs
@@ -1,9 +1,11 @@
 use syn;
+use proc_macro2::TokenStream;
 use syn::spanned::Spanned;
 use syn::parse::Error;
+use quote::quote;
 
 use crate::util::ident_to_type;
-use crate::builder_attr::FieldBuilderAttr;
+use crate::util::{expr_to_single_string, path_to_single_string};
 
 #[derive(Debug)]
 pub struct FieldInfo<'a> {
@@ -48,5 +50,108 @@ impl<'a> FieldInfo<'a> {
             paren_token: Default::default(),
             elems: types,
         }.into()
+    }
+}
+
+#[derive(Debug, Default)]
+pub struct FieldBuilderAttr {
+    pub doc: Option<syn::Expr>,
+    pub exclude: bool,
+    pub default: Option<syn::Expr>,
+}
+
+impl FieldBuilderAttr {
+    pub fn new(attrs: &[syn::Attribute]) -> Result<FieldBuilderAttr, Error> {
+        let mut result = FieldBuilderAttr::default();
+        let mut exclude_tts = None;
+        for attr in attrs {
+            if path_to_single_string(&attr.path).as_ref().map(|s| &**s) != Some("builder") {
+                continue;
+            }
+
+            if attr.tts.is_empty() {
+                continue;
+            }
+
+            let as_expr: syn::Expr = syn::parse2(attr.tts.clone())?;
+            match as_expr {
+                syn::Expr::Paren(body) => {
+                    result.apply_meta(*body.expr)?;
+                }
+                syn::Expr::Tuple(body) => {
+                    for expr in body.elems.into_iter() {
+                        result.apply_meta(expr)?;
+                    }
+                }
+                _ => {
+                    return Err(Error::new_spanned(attr.tts.clone(), "Expected (<...>)"));
+                }
+            }
+            // Stash its span for later (we don’t yet know if it’ll be an error)
+            if result.exclude && exclude_tts.is_none() {
+                exclude_tts = Some(attr.tts.clone());
+            }
+        }
+
+        if result.exclude && result.default.is_none() {
+            return Err(Error::new_spanned(
+                exclude_tts.unwrap(),
+                "#[builder(exclude)] must be accompanied by default or default_code",
+            ));
+        }
+
+        Ok(result)
+    }
+
+    fn apply_meta(&mut self, expr: syn::Expr) -> Result<(), Error> {
+        match expr {
+            syn::Expr::Assign(assign) => {
+                let name = expr_to_single_string(&assign.left).ok_or_else(
+                    || Error::new_spanned(&assign.left, "Expected identifier"))?;
+                match name.as_str() {
+                    "doc" => {
+                        self.doc = Some(*assign.right);
+                        Ok(())
+                    }
+                    "default" => {
+                        self.default = Some(*assign.right);
+                        Ok(())
+                    }
+                    "default_code" => {
+                        if let syn::Expr::Lit(syn::ExprLit{lit: syn::Lit::Str(code), ..}) = *assign.right {
+                            use std::str::FromStr;
+                            let tokenized_code = TokenStream::from_str(&code.value())?;
+                            self.default = Some(syn::parse(tokenized_code.into()).map_err(|e| Error::new_spanned(code, format!("{}", e)))?);
+                        } else {
+                            return Err(Error::new_spanned(assign.right, "Expected string"));
+                        }
+                        Ok(())
+                    },
+                    _ => {
+                        Err(Error::new_spanned(&assign, format!("Unknown parameter {:?}", name)))
+                    }
+                }
+            }
+            syn::Expr::Path(path) => {
+                let name = path_to_single_string(&path.path).ok_or_else(
+                    || Error::new_spanned(&path, "Expected identifier"))?;
+                match name.as_str() {
+                    "exclude" => {
+                        self.exclude = true;
+                        Ok(())
+                    }
+                    "default" => {
+                        self.default = Some(syn::parse(quote!(Default::default()).into()).unwrap());
+                        Ok(())
+                    }
+                    _ => {
+                        Err(Error::new_spanned(&path, format!("Unknown parameter {:?}", name)))
+                    }
+                }
+            }
+            _ => {
+                Err(Error::new_spanned(expr, "Expected (<...>=<...>)"))
+            }
+        }
     }
 }
