+++ src/struct_info.rs
@@ -5,8 +5,8 @@ use syn::parse::Error;
 use quote::quote;
 
 use crate::field_info::FieldInfo;
-use crate::builder_attr::TypeBuilderAttr;
 use crate::util::{empty_type, make_punctuated_single, modify_types_generics_hack};
+use crate::util::{expr_to_single_string, path_to_single_string};
 
 #[derive(Debug)]
 pub struct StructInfo<'a> {
@@ -323,5 +323,105 @@ impl<'a> StructInfo<'a> {
                 }
             }
         ).into()
+    }
+}
+
+#[derive(Debug, Default)]
+pub struct TypeBuilderAttr {
+    /// Whether to show docs for the `TypeBuilder` type (rather than hiding them).
+    pub doc: bool,
+
+    /// The name of the builder type.
+    pub name: Option<syn::Expr>,
+
+    /// Docs on the `Type::builder()` method.
+    pub builder_method_doc: Option<syn::Expr>,
+
+    /// Docs on the `TypeBuilder` type. Specifying this implies `doc`, but you can just specify
+    /// `doc` instead and a default value will be filled in here.
+    pub builder_type_doc: Option<syn::Expr>,
+
+    /// Docs on the `TypeBuilder.build()` method. Specifying this implies `doc`, but you can just
+    /// specify `doc` instead and a default value will be filled in here.
+    pub build_method_doc: Option<syn::Expr>,
+}
+
+impl TypeBuilderAttr {
+    pub fn new(attrs: &[syn::Attribute]) -> Result<TypeBuilderAttr, Error> {
+        let mut result = TypeBuilderAttr::default();
+        for attr in attrs {
+            if path_to_single_string(&attr.path).as_ref().map(|s| &**s) != Some("builder") {
+                continue;
+            }
+
+            if attr.tts.is_empty() {
+                continue;
+            }
+            let as_expr: syn::Expr = syn::parse2(attr.tts.clone())?;
+
+            match as_expr {
+                syn::Expr::Paren(body) => {
+                    result.apply_meta(*body.expr)?;
+                }
+                syn::Expr::Tuple(body) => {
+                    for expr in body.elems.into_iter() {
+                        result.apply_meta(expr)?;
+                    }
+                }
+                _ => {
+                    return Err(Error::new_spanned(attr.tts.clone(), "Expected (<...>)"));
+                }
+            }
+        }
+
+        Ok(result)
+    }
+
+    fn apply_meta(&mut self, expr: syn::Expr) -> Result<(), Error> {
+        match expr {
+            syn::Expr::Assign(assign) => {
+                let name = expr_to_single_string(&assign.left).ok_or_else(
+                    || Error::new_spanned(&assign.left, "Expected identifier"))?;
+                match name.as_str() {
+                    "name" => {
+                        self.name = Some(*assign.right);
+                        Ok(())
+                    }
+                    "builder_method_doc" => {
+                        self.builder_method_doc = Some(*assign.right);
+                        Ok(())
+                    }
+                    "builder_type_doc" => {
+                        self.builder_type_doc = Some(*assign.right);
+                        self.doc = true;
+                        Ok(())
+                    }
+                    "build_method_doc" => {
+                        self.build_method_doc = Some(*assign.right);
+                        self.doc = true;
+                        Ok(())
+                    }
+                    _ => {
+                        Err(Error::new_spanned(&assign, format!("Unknown parameter {:?}", name)))
+                    }
+                }
+            }
+            syn::Expr::Path(path) => {
+                let name = path_to_single_string(&path.path).ok_or_else(
+                    || Error::new_spanned(&path, "Expected identifier"))?;
+                match name.as_str() {
+                    "doc" => {
+                        self.doc = true;
+                        Ok(())
+                    }
+                    _ => {
+                        Err(Error::new_spanned(&path, format!("Unknown parameter {:?}", name)))
+                    }
+                }
+            }
+            _ => {
+                Err(Error::new_spanned(expr, "Expected (<...>=<...>)"))
+            }
+        }
     }
 }
